feat: :lipstick: added frost effect to entire navbar, relocated guestbook entries to seperate file and changed colors in meta component
