+++ public/site.webmanifest
@@ -2,8 +2,8 @@
     "name": "Dhruv Gursahani",
     "description": "Software developer and dual student",
     "start_url": "/",
-    "background_color": "#fff",
-    "theme_color": "#000",
+    "background_color": "#f9fafb",
+    "theme_color": "#222222",
     "short_name": "Dhruv",
     "display": "standalone",
     "orientation": "portrait",
