+++ styles/globals.css
@@ -222,8 +222,8 @@ pre {
   @apply my-0;
 }
 
-.glass-bg {
-  @apply bg-gray-50 dark:bg-gray-800 md:bg-opacity-50 md:dark:bg-opacity-60 backdrop-blur-lg backdrop-saturate-150;
+.frost-effect {
+  @apply bg-gray-50 dark:bg-gray-800 bg-opacity-50 dark:bg-opacity-60 backdrop-blur-lg backdrop-saturate-150;
 }
 
 .link-underline {
